+++ react/__mocks__/react-content-loader.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ContentLoad = ({ children }) => (
+  <div>
+    {children}
+  </div>
+)
+
+export default ContentLoad