+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Quick fix on `AutocompleteInput` to remove warnings.
 ### Added
 - Base settings for tests.
+- Snapshot tests for ProductPrice.
 
 ## [3.17.0] - 2019-02-15
 ### Added